Persistence manager for pickle, rename file per convention.
