+++ crates/holochain_persistence_file/src/txn/mod.rs
@@ -1,2 +1,2 @@
-mod txn;
-pub use txn::*;
+mod file;
+pub use file::*;
