+++ crates/holochain_persistence_pickle/src/lib.rs
@@ -11,3 +11,4 @@ extern crate test;
 
 pub mod cas;
 pub mod eav;
+pub mod txn;
